+++ test/sentry/config_traces_sampler_test.exs
@@ -1,5 +1,5 @@
 defmodule Sentry.ConfigTracesSamplerTest do
-  use ExUnit.Case, async: true
+  use ExUnit.Case, async: false
 
   import Sentry.TestHelpers