+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 11.0.3
+
+#### Various improvements
+
+- Allow any version of opentelemetry deps and verify minimum versions internally - this makes it possible to use `sentry` *with tracing disabled* along with older versions of opentelemetry deps ([#931](https://github.com/getsentry/sentry-elixir/pull/931))
+
+## 11.0.2
+
+### Bug fixes
+
+- Deeply nested spans are handled now when building up traces in `SpanProcessor` ([#924](https://github.com/getsentry/sentry-elixir/pull/924))
+
+#### Various improvements
+
+- Span's attributes no longer include `db.url: "ecto:"` entries as they are now filtered out ([#925](https://github.com/getsentry/sentry-elixir/pull/925))
+
 ## 11.0.1
 
 #### Various improvements